requirements copy.txt
google-adk>=1.3.0
google-cloud-aiplatform>=1.35.0
google-cloud-storage>=2.10.0
google-cloud-functions>=1.13.0
google-cloud-pubsub>=2.18.0
pydantic>=2.0.0
asyncio>=3.4.3
typing-extensions>=4.0.0 
pillow 
statistics
logging 
datetime
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai>=0.3.0
